[wip] Day 16 part 1
Backtracking isn't working properly, and we get stuck in a loop.
Trying to backtrack to the last junction, so we can do the next
available turn.

It should complete a depth-first search of the graph so that eventually
every possible route from S->E has been visited, and so we have the
lowest cost.